requirements.txt
tensorflow-cpu==2.13.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.2
librosa==0.10.1
soundfile==0.12.1
speechrecognition==3.10.1
streamlit==1.28.0
matplotlib==3.7.5
seaborn==0.13.0
pillow==10.0.0
